corrected routing errors and js errors in public folder

+++ controllers/api/comment-routes.js
@@ -29,22 +29,24 @@ router.post('/', (req, res) => {
   
 
 router.delete('/:id', (req, res) => {
-  Comment.destroy({
-    where: {
-      id: req.params.id
-    }
-  })
-    .then(dbCommentData => {
-      if (!dbCommentData) {
+  if (req.session) {
+    Comment.destroy({
+      where: {
+        id: req.params.id
+      }
+    })
+      .then(dbCommentData => {
+       if (!dbCommentData) {
         res.status(404).json({ message: 'No comment found with this id!' });
         return;
       }
       res.json(dbCommentData);
-    })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json(err);
-    });
+      })
+      .catch(err => {
+        console.log(err);
+        res.status(500).json(err);
+      });
+  }
 });
 
 module.exports = router;
